narayaneeyam/requirements.txt
# django
django==4.2.17
django-hashids==0.7.0
djangorestframework==3.15.2
django-phonenumber-field[phonenumberslite]==8.0.0
django-pg-jsonschema==0.1.2

# CORS
django-cors-headers==4.6.0

# local development
psycopg2-binary==2.9.10
python-dotenv==1.0.1
setuptools==75.6.0

# swagger
drf-spectacular[sidecar]==0.27.2
whitenoise==6.8.2

# tailwind
pytailwindcss==0.2.0

# other
pytz==2024.2